Rename test/lib/buildwatcher to test/lib/buildrunner
