+++ pyproject.toml
@@ -123,7 +123,7 @@ gbcli = "gbcli.cli:gbcli"
 gb = "gbcli.cli:gbcli"
 llmbuild = "gbcli.cli:gbcli"
 # gbtest: run a single buildtest.yaml via the pytest harness
-gbtest = "lib.buildwatcher.gbtest:main"
+gbtest = "lib.buildrunner.gbtest:main"
 llmb = "gbcli.cli:gbcli"
 lamb = "gbcli.cli:gbcli"
