+++ test/integration/ibm/buildrunner/k8s/test_buildrunner_1step_gpu.py
@@ -16,7 +16,7 @@ greppable rather than buried in auto-discovery convention.
 from pathlib import Path
 
 import pytest
-from lib.buildwatcher.buildtest import (
+from lib.buildrunner.buildtest import (
     AbstractYamlBuildRunnerTest,
     get_test_data_dir_for,
 )
