+++ test/integration/ibm/buildrunner/k8s/test_buildrunner_retry.py
@@ -3,13 +3,13 @@ from time import sleep, time
 from typing import Optional, Self
 
 import pytest
-from lib.buildwatcher.buildtest import (
+from lib.buildrunner.buildtest import (
     AbstractBuildTest,
     BuildTestSpecification,
     ClassTestedEnum,
     ExpectedTarget,
 )
-from lib.buildwatcher.utils import ExceptionRaisingThread
+from lib.buildrunner.utils import ExceptionRaisingThread
 from lib.constants import (
     GBTEST_SPACE_NAME,
     GBTEST_USER_NAME,
