+++ test/integration/ibm/buildrunnerjob/k8s/test_buildrunnerjob.py
@@ -1,7 +1,7 @@
 import os
 
 import pytest
-from lib.buildwatcher.buildtest import (
+from lib.buildrunner.buildtest import (
     AbstractBuildTest,
     BuildTestSpecification,
     ClassTestedEnum,
